Adds handing of z-channel only passes
This will look if the EXR channel only has a z sub-channel, and if so will assign the z-channel to the R, G, B, and Z channels of the Loader.

There is still an issue in Prism where it will not pass these channels to the Fusion plugin.

See issue #117