Added ability to skip Wireless_OUT reposition #138
This checks if the Wireless_OUT has been moved away from the Loader stack by the user.  This allows the user to reposition the Wireless into the comp flow if desired and not have its position affected by other re-sorts.

This checks the distance of the _OUT to the _IN, and if below the threshold it will reposition.