Fixed Issue from Fixed issue #127
The original fix passed self.statemanger to the callback in the Libraries plugin.  But found on one machine there is an issue that I think comes from the order of plugin's loading.

What I saw was that if a user opens the SM and imports a 3d item, then closes the SM when they reopen the SM it does not load correctly.

This fix just disables the callback functions to the Libraries plugin until more investigation.  This means when importing 3d it will always open the ProjectBrowser which is not a big deal.