Initial implementation of new type of Tool Data #152
This changes the structure of the data saved in the tools.  This is to allow rebuilding of the CompDB and for future use.  This separates the methods into two:  the first configures the tool settings and the second saves the data in a dict in the tool's custom data.

Right now these are working:
- USD Import
- 3d Import
-USD Mat import
-ShotCam import
- Legacy Import (mostly)

These are not working and need more work:
- Image Import loses positioning
- There is still an error with updating Legacy3d sometimes.

Other changes:
- changed some logging to debug to stop unneeded red consoles
- changed name of the data saved in the tool from "Prism_dbData" to "Prism_ToolData"
- changed method to "configureTool()"
- changed some calls to te above methods
- changed the metadata from "3dFilepath" to "object3dFilepath" since Fusion doesnt like having a leading number
-