Fixes #156 Error when updating Legancy3d
This adds a check if there is a valid connected tool before trying to reconnect.

This comes from if the Root tool has no other tools connected to its output but is being viewed in a viewer, then it will have a connection but it is not  tool.

An enhancement could be to also re-show the tool in the viewer after updating, but that is not implemnented here yet.