Added functions for AOV / Channel thumbnail display #158
Added ability to display an image thumbnail for each AOV / Channel in the tree list.  But this can slow down performance with large images or video files.

This tries to use the Prism thumb (in dir  _thumb) to aid in performance, but still slow with large video files.

This has an option in Settings-->DCC-->Fusion to enable or disable the thumbnails.

Also removed the now unused Color Brightness from Settings.