Fixed Video Import #182 and added Frame Range to UI
This fixes an issue with some video file imports.  Should have seen it before, but it is a lowercase issue.  Just added a ".lower()" method.

While in there, added the display of the framerange for the images in the AOV/Channel list UI.  This will show the framerange for each file - so each AOV will have their own framerange, and images without an AOV will have it displayed next to the identifier at the top.