Fixed ability to import images with same MediaID from different shots #187
This updates the keys that are checked to compare import items to include Asset/Seq/Shot.  So if a user imports an image with the same identifier from another shot, it will be treated as a separate import.

This seems to be working, but further testing needed.

This does break the task selecting and coloring a bit.  The selecting/coloring is still relying on the MediaID, so will select all the tools with the same MediaID.  That needs to be fixed.