Fixed issue from the import comparision in #187
With the code from the original fix to check for the same MediaID's from different entities, there were sometimes false positives.  This came from the fact that there was a discrepancy between the naming of "identifier" and "mediaId".

This fixes that, so now compares correctly.

This also has a couple of small code clean-ups.