Changed AOV list selecting and re-ordered code
Changed the UI behavior of the AOV list to use checkboxes instead of selection highlighting.  This adds checkboxes to each row and allows the user to check which "passes" are desired for a given identifier.

The checkbox status is still still not saving its state in the Comp, that is still WIP.

Also went ahead and reordered the code to group together the methods, and added header comments.