Added UI to the DCC Settings for Sorting #200
This adds a combo to the Fusion DCC settings to allow the user to select what Loader sorting is desired.

The three options are:

"Sorting & Wireless": acts the way we have been doing it.
"Sorting Only": will add and sort the Loaders, but does not add wireless tools
"Disabled": will not sort or add wireless.  The Loader will be added to the flow the native "Fusion way".