Added multi-selection importing #196 and Creating State from Settings #159
This adds functionality to be able to multi-select Media Identifiers and right-click to import.

If there are multiple items, it will actually call the addImportState() from the main Prims_Fusion_Functions which in turn calls the StateManager to create the State from passed setting data.

This in turn implements code to be able to create an Image Import state from the main Fusion_Functions with passed settings data.