Minimizes the StateManager for a Local Render #219
This will minimize the State Manager when a Local Render is triggered. This allows the user to see the render, and the resulting Fusion Popup.

As of now, the StateManager is not restored since I did not find a way (callback) to trigger the restore after the Prism "Finshed" popup is closed. So in my opinion it would then re-cover the popup (but the code is there and commented out).

Again this just minimizes the SM, since using the .close() method then prevents reopening the SM for some reason.