Added Creating Import State from Orphaned Loader(s) in the Comp #222
This adds functionality to scan the Comp for all Prism Loaders.  It then compares the Loader's State UUID to all the existing State UIDs in the State Manager.  If it is not found, this can create a new State using the data stored in the Loader.

This also adds a DCC user setting to control this state creation:  "Auto", "Prompt", and "Disabled".