Makes Import Ltest use Existing Checked AOVs #230
Previously the Import Latest button uses the ImportAll() call. This means that if there are selected AOVs already in the Comp, using the Import Latest will not respect the selection and will import all AOVs.

It is a design choice, but it would prob be better to respect the selection. The downside that if the newer version contains more AOV's than the previous, they will not automatically be imported - but that may be a less-likely case.