Fixed #234 - Not an issue with actual importing, but just status coloring
Soo.  I do not think there was anything wrong with the versioning, there was an error in my test images.

But there was a small error where the Status Color may not be updated in some instances.  That seems fixed now.

While I was in there, I also added logic to keep the checked AOVs states after a version is chosen with the Select Version button